Rename the errors module to exceptions
