+++ press/legacy_publishing/module.py
@@ -1,8 +1,10 @@
 from pyramid.threadlocal import get_current_request
 from sqlalchemy.sql import text
 
-from .utils import replace_id_and_version, produce_hashes_from_filepath
-from ..errors import StaleVersion, Unchanged
+from press.utils import produce_hashes_from_filepath
+from press.exceptions import StaleVersion, Unchanged
+from .utils import replace_id_and_version
+
 
 __all__ = (
     'publish_legacy_page',
